requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
# RAG
langchain
langchain-openai
langchain-community
qdrant-client
# Utilities
httpx
pydantic
